Add DayTitles file to hold days of the week text
The days of the week titles that are arranged in a ColumnLayout were
moved to a separate file and referenced in main.qml using the name of
the file, DayTitles.

The days of the week text qml code was moved to a separate file to
improve code separation and increase readability of main.qml.