Change enterInputRect to fill the Calendar
The size of enterInputRect was changed to fit the size of the Calendar
and the anchors of dayRect were changed to center in parent. This
makes the dayRect stay centered in the Calendar.

Changing enterInputRect to fill Calendar allows for dayRect to always
remain centered, regardless of what size the window is. When attempting
to center the rectangle in the Calendar by accessing the id the
alignment would not change, and this change allows for the intended
alignment.