Define a button that fits text to a rectangle
Define an element using NavigationButton.qml that contains a rectangle
with fitted text that can be reused for the navigation buttons in the
user input screen. The size and color of the rectangle can be changed
through properties and the content of the text can be changed through a
property.

Defining a custom button reduces code duplication and makes the code
easier to read.