Add case when the user doesn't select a button
Add case to the setRadioButton() function for when the user does not
select a radio button and add the function setRadioButtonsFalse to set
the checked value for all radio buttons to false.

The case when the user did not select a radio button, but entered other
values was not caught and would lead to an error when nothing is
inserted or retrieved for the radio buttons.