Add logic to update date, use next day button
The date that is displayed on the UserInputScreen is updated every time
that the visibility of the UserInputScreen changes. The logic for the
next day button was implemented so that the database is updated and
then the values of the elements on the UserInputScreen are updated using
an sql query to advance to the next day.

Updating the date fixes a bug where the date would be incorrect after
advancing days and selecting a new date after backtracking to the
calendar view. The database logic was added to the next day button so
that the button advances the view to the next day.