Fix bug where radio buttons were not set properly
A bug occurred previously when the radio buttons were set using the
setRadioButton() function. The latest radio button in the order would
be the only radio button displayed after it is set, this was caused by
multiple buttons having a checkedVal that is true and the latest value
that was true was the value that was displayed. This was fixed by
setting all other radio buttons to false when a value is set.

This bug prevented the mood selection to be displayed incorrectly and
is a key functionality for the program.